update to 9.1.0

+++ TMessagesProj/src/main/java/org/telegram/ui/Components/RLottieDrawable.java
@@ -21,7 +21,6 @@ import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.HapticFeedbackConstants;
 import android.view.View;
 
@@ -1243,7 +1242,7 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
     @Override
     public void prepareForGenerateCache() {
         generateCacheNativePtr = create(args.file.toString(), args.json, width, height, new int[3], false, args.colorReplacement, false, args.fitzModifier);
-        if (generateCacheNativePtr == 0) {
+        if (generateCacheNativePtr == 0 && file != null) {
             file.delete();
         }
     }
